Use the new FileSystemManager config resolver
